vertical loss
