+++ federatedscope/vertical_fl/loss/__init__.py
@@ -1,0 +1,5 @@
+from federatedscope.vertical_fl.loss.binary_cls import BinaryClsLoss
+from federatedscope.vertical_fl.loss.regression import RegressionMAELoss, \
+    RegressionMSELoss
+
+__all__ = ['BinaryClsLoss', 'RegressionMSELoss', 'RegressionMAELoss']
